+++ packages/terra-date-picker/CHANGELOG.md
@@ -3,11 +3,21 @@ Changelog
 
 Unreleased
 ----------
+### Changed
+* Update tests for terra-toolkit v5 and terra-dev-site v5 changes
 
-4.1.0 - (March 15, 2019)
+4.2.0 - (May 21, 2019)
+------------------
+### Changed
+* Minor dependency update
+
+4.1.0 - (May 15, 2019)
 ------------------
 ### Changed
 * Resolved eslint error in example files
+
+### Added
+* Added DatePickerField
 
 4.0.0 - (May 13, 2019)
 ------------------
@@ -15,6 +25,8 @@ Unreleased
 * Removed `releaseFocus` prop
 * Removed `requestFocus` prop
 * Removed `onInputFocus` prop in favor of the added `onFocus` prop.
+
+### Changed
 * Changed the parameter in filterDate prop from a moment object to an ISO string.
 * Honor min/max range when only either minDate or maxDate is set.
 * Changed the onBlur event to get triggered only when the entire date picker component loses focus.